refactor(compartment-mapper): Rename assemble.js to link.js
