+++ packages/compartment-mapper/src/import-archive.js
@@ -15,7 +15,7 @@
 /** @typedef {import('./types.js').ExecuteOptions} ExecuteOptions */
 
 import { readZip } from '@endo/zip';
-import { link } from './assemble.js';
+import { link } from './link.js';
 import { parsePreCjs } from './parse-pre-cjs.js';
 import { parseJson } from './parse-json.js';
 import { parsePreMjs } from './parse-pre-mjs.js';
