+++ packages/endo/bin/endo.js
@@ -2,6 +2,7 @@
 /* global process */
 (async () => {
   const fs = await import('fs');
+  const crypto = await import('crypto');
   const { main } = await import('../src/cli.js');
-  main(process, { fs: fs.promises });
+  main(process, { fs, crypto });
 })();
